rename helper file of table of contents
