+++ src/helpers/tableContents.helper.ts
